+++ test/OpenTelemetry.Tests/Metrics/MeterProviderSdkTest.cs
@@ -1,0 +1,48 @@
+// <copyright file="MeterProviderSdkTest.cs" company="OpenTelemetry Authors">
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+using Xunit;
+
+namespace OpenTelemetry.Metrics.Tests;
+
+public class MeterProviderSdkTest
+{
+    [Fact]
+    public void BuilderTypeDoesNotChangeTest()
+    {
+        var originalBuilder = Sdk.CreateMeterProviderBuilder();
+        var currentBuilder = originalBuilder;
+
+        var deferredBuilder = currentBuilder as IDeferredMeterProviderBuilder;
+        Assert.NotNull(deferredBuilder);
+
+        currentBuilder = deferredBuilder.Configure((sp, innerBuilder) => { });
+        Assert.True(ReferenceEquals(originalBuilder, currentBuilder));
+
+        currentBuilder = currentBuilder.ConfigureServices(s => { });
+        Assert.True(ReferenceEquals(originalBuilder, currentBuilder));
+
+        currentBuilder = currentBuilder.AddInstrumentation(() => new object());
+        Assert.True(ReferenceEquals(originalBuilder, currentBuilder));
+
+        currentBuilder = currentBuilder.AddMeter("MySource");
+        Assert.True(ReferenceEquals(originalBuilder, currentBuilder));
+
+        using var provider = currentBuilder.Build();
+
+        Assert.NotNull(provider);
+    }
+}